+++ Assets/Scripts/Genetics/GeneticsUtils.cs
@@ -3,11 +3,11 @@ using System.Collections;
 
 public class GeneticsUtils {
 
-	public static byte[] mutate (byte[] c) {
+	public static float[] mutate (float[] c) {
 		return c;
 	}
 	
-	public static byte[] crossover (byte[] c1, byte[] c2) {
+	public static float[] crossover (float[] c1, float[] c2) {
 		return c1;
 	}
